requirements.txt
praw==5.4.0
prawcore==0.14.0
vcrpy==1.12.0
python-dotenv==0.8.2
Jinja2==2.10
raven==6.9.0
